Move and rename formatter classes
